refactor: rename components and update import paths for consistency
