+++ src/pages/Chatbox.tsx
@@ -24,10 +24,9 @@ import { models } from "@/types/models";
 import { AddImage } from "@/addImage";
 import { ListAPIs } from "@/listAPIs";
 import { ListToolsTempaltes } from "@/listToolsTemplates";
-import { autoHeight } from "@/textarea";
+import { autoHeight } from "@/utils/textAreaHelp";
 import Templates from "@/components/Templates";
 import VersionHint from "@/components/VersionHint";
-import StatusBar from "@/components/StatusBar";
 import WhisperButton from "@/components/WhisperButton";
 import AddToolMsg from "./AddToolMsg";
 import { Textarea } from "@/components/ui/textarea";
